requirements.txt
# CKSEARCH - OSINT Intelligence Tool
# Dependencies

# HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.1.0
fake-useragent>=1.4.0

# DNS & Network
dnspython>=2.4.0
python-whois>=0.8.0
socket

# Phone Analysis
phonenumbers>=8.13.0

# Image Processing
Pillow>=10.0.0
exifread>=3.0.0

# CLI & UI
rich>=13.7.0
colorama>=0.4.6
pyfiglet>=1.0.2
tqdm>=4.66.0
questionary>=2.0.0

# Report Generation
jinja2>=3.1.2

# Data Processing
validators>=0.22.0

# Cryptography (for hashing)
hashlib

# Optional: Tor support
# PySocks>=1.7.1
